+++ modules/canbus/vehicle/lincoln/protocol/throttle_62.cc
@@ -71,15 +71,15 @@ void Throttle62::set_pedal_p(uint8_t *data, double pedal) {
   // and a rough mapping
   pedal /= 100.0;
   pedal = ProtocolData::BoundedValue(0.0, 1.0, pedal);
-  int32_t x = pedal / 1.52590218966964e-05;
+  int32_t x = static_cast<int32_t>(pedal / 1.52590218966964e-05);
   std::uint8_t t = 0;
 
-  t = x & 0xFF;
+  t = static_cast<uint8_t>(x & 0xFF);
   Byte frame_low(data + 0);
   frame_low.set_value(t, 0, 8);
 
   x >>= 8;
-  t = x & 0xFF;
+  t = static_cast<uint8_t>(x & 0xFF);
   Byte frame_high(data + 1);
   frame_high.set_value(t, 0, 8);
 }
@@ -114,7 +114,7 @@ void Throttle62::set_ignore_driver_override_p(uint8_t *bytes, bool ignore) {
 void Throttle62::set_watchdog_counter_p(uint8_t *data, int32_t count) {
   count = ProtocolData::BoundedValue(0, 255, count);
   Byte frame(data + 7);
-  frame.set_value(count, 0, 8);
+  frame.set_value(static_cast<uint8_t>(count), 0, 8);
 }
 
 }  // namespace lincoln
